ml-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
pillow>=10.1.0
numpy>=1.24.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
librosa>=0.10.1
scipy>=1.11.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
qdrant-client>=1.6.0
open-clip-torch>=2.20.0  # Alternative to openai-clip for better compatibility
soundfile>=0.12.1
# Additional dependencies for enhanced functionality
matplotlib>=3.7.0  # For audio visualization (optional)
seaborn>=0.12.0   # For data visualization (optional)
pandas>=2.0.0     # For data manipulation (optional)
asyncio-mqtt>=0.13.0  # For potential real-time features (optional)
redis>=5.0.0      # For caching (optional)
